added cantilever to montage and preview

+++ montage.sh
@@ -10,6 +10,12 @@ then
     echo "unrecognized input"
     exit 1
 fi
+read -r -e -p "create new miscellaneous output? (y/n): " create_miscellaneous
+if ! [[ "$create_miscellaneous" == "y" || "$create_miscellaneous" == "Y" || "$create_miscellaneous" == "n" || "$create_miscellaneous" == "N" ]]
+then
+    echo "unrecognized input"
+    exit 1
+fi
 mkdir -p tmp/montage
 rm -rf tmp/montage/*
 if [[ "$create_bridges" == "y" || "$create_bridges" == "Y" ]]
@@ -19,6 +25,10 @@ fi
 if [[ "$create_pendulums" == "y" || "$create_pendulums" == "Y" ]]
 then
     ./pendulums.sh
+fi
+if [[ "$create_miscellaneous" == "y" || "$create_miscellaneous" == "Y" ]]
+then
+    ./miscellaneous.sh
 fi
 mkdir -p tmp/montage/bridges/warren
 echo "\
@@ -129,6 +139,18 @@ ffmpeg \
     -map "[v]" \
     -y tmp/montage/pendulums/decuplependulum/video.mp4 \
     -loglevel error
+mkdir -p tmp/montage/miscellaneous/cantilever
+echo "subtitles=1
+center=(0.0 -0.4) lineheight=0.05 text=Cantilever
+" | ./bin/subtitles tmp/montage/miscellaneous/cantilever fsize=1920x1080
+ffmpeg \
+    -i tmp/miscellaneous/cantilever/video.mp4 \
+    -i tmp/miscellaneous/cantilever/fdiagram.mp4 \
+    -i tmp/montage/miscellaneous/cantilever/subtitles.png \
+    -filter_complex "[1:v]scale=640:360[y];[0:v][y]overlay[z];[z][2:v]overlay[v]" \
+    -map "[v]" \
+    -y tmp/montage/miscellaneous/cantilever/video.mp4 \
+    -loglevel error
 ffmpeg \
     -i tmp/montage/bridges/warren/video.mp4 \
     -i tmp/montage/bridges/pratt/video.mp4 \
@@ -139,9 +161,11 @@ ffmpeg \
     -i tmp/montage/pendulums/doublependulum/video.mp4 \
     -i tmp/montage/pendulums/doublependulumroller/video.mp4 \
     -i tmp/montage/pendulums/decuplependulum/video.mp4 \
-    -filter_complex "[0:v][1:v][2:v][3:v][4:v][5:v][6:v][7:v][8:v]concat=n=9:v=1[v]" \
+    -i tmp/montage/miscellaneous/cantilever/video.mp4 \
+    -filter_complex "[0:v][1:v][2:v][3:v][4:v][5:v][6:v][7:v][8:v][9:v]concat=n=10:v=1[v]" \
     -map "[v]" \
     -y tmp/montage/video.mp4 \
     -loglevel error
 rm -rf tmp/montage/bridges
 rm -rf tmp/montage/pendulums
+rm -rf tmp/montage/miscellaneous
